fix(Select.stories): implement controller for history within a form
